Fix crash after winning/dying on a continued game
TPlayer's FKills field was being registered with the Lua system as nil,
i.e. before being properly initialized, if and only if the player was
resuming a saved game.  The game would then crash consistently on death
or a win when trying to award medals, specifically when checking the
kill table for the John Carmack kill; consulting this nil kill table
resulted in an EAccessViolation.

This commit fixes the above crash by registering the player with the Lua
system after ALL of its fields are valid when continuing a game.

Dying or winning a game in a single play session does not have this bug.