Unnerf and rework assault rifle assembly
The old assault rifle assembly was possibly the only weapon assembly in
the game that _reduced_ damage-per-second.  This commit changes the
assembly to be more generally useful.

The old assault rifle assembly:

 * 50% reload time
 * +1 damage dice
 * -1 damage sides (old)
 * x2 shot cost (old)

The new assault rifle assembly:

 * 50% reload time
 * +1 damage dice
 * 65% ammo capacity (new)

The new assault rifle assembly halves ammo consumed per fired burst and
reduces the downsides of reloading, effectively making it a sort of
Ammochain-lite for non-Ammochain builds.